Add files via upload
When logged in, the option to log in turns into logout. Clicking logout should log you out of your account, but not disconnect from the database. 
Now, when an option is selected to search by which does not need the text field, the text field will be disabled on the search dialog. Feel free to let me know if I've missed some that don't use it.
Also, the history option only shows up if you are logged in. All of this functionality can be added to any other search, just need to let me know if others need it.